services/storage-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
aioboto3==12.0.0
zstandard==0.22.0
cryptography==41.0.7
kafka-python==2.0.2
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-client==0.19.0
elasticsearch[async]==8.11.0
clickhouse-driver==0.2.6
celery==5.3.4
sqlalchemy[asyncio]
email-validator==2.1.0
requests==2.31.0
aiokafka==0.10.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
aiocache[redis]==0.12.2
psutil==5.9.6
alembic==1.13.1